+++ combat/modifier/ModifierMove.java
@@ -6,7 +6,7 @@ import field.Field;
 
 public class ModifierMove extends Modifier {
 
-    public ModifierMove(MessageModifier message, Float modifier, Function<Field, Boolean> check) {
+    public ModifierMove(MessageModifier message, double modifier, Function<Field, Boolean> check) {
         super(message, modifier);
         this.moveCheck = check;
     }
